Works

+++ sandbox/src/data_compressor/paa/compress_paa.py
@@ -18,7 +18,7 @@ class CompressPAA(Compressor):
     if data_size < 2:
       self.compressed_data = self.original_data[:]
       return
-    chunk_count = data_size * self.config['compress_ratio']
+    chunk_count = int(data_size * self.config['compress_ratio'])
     x_first = self.original_data[0].timestamp
     x_last = self.original_data[data_size - 1].timestamp
     chunk_size = (x_last - x_first) / chunk_count
